Fix rendering of status paths in indicator
The status (unknown, alert, charging) paths appeared to be rendering underneath
the percentage path, and changing the fill of these paths made them disappear.
According to [section 3.3 of the the SVG specification](https://www.w3.org/TR/SVG11/render.html#RenderingOrder),
"Subsequent elements are painted on top of previously painted elements," which
explains the aforementioned behavior. Moving the status paths to the end of the
SVG fixes the issue.