Scale to fit container and use transform
Change the indicator element to scale to fit its container instead of
using an explicit height/width of 48px. This requires setting its height
and width in the page that is using it, but provides flexibility to the
consumer.

Change the method of setting the charge path width from editing the path
itself to applying a transform on the path. This is a slightly cleaner
approach that may be more compatible with transition animations in the
future.